Create App folder & relocate App files
